Move mysensors base entity to separate module
