+++ homeassistant/components/mysensors/entity.py
